move decorator to util.py
